requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-bitcoinlib>=0.11.0
pytest>=6.2.5
requests>=2.26.0
python-dotenv>=0.19.0
python-json-logger>=2.0.7
pytest-cov>=2.12.0 